Refactor Pascal parser into smaller files
This change refactors the Pascal parser by splitting the code from a single large file (`pascal_parser.c`) into smaller, more focused files. This improves the organization and readability of the code.

The following changes were made:
- Created new files for different parts of the parser:
  - `pascal_expression.c` and `pascal_expression.h` for parsing expressions.
  - `pascal_statement.c` and `pascal_statement.h` for parsing statements.
  - `pascal_declaration.c` and `pascal_declaration.h` for parsing declarations.
  - `pascal_type.c` and `pascal_type.h` for parsing types.
  - `pascal_keywords.c` and `pascal_keywords.h` for managing keywords.
- Moved the relevant code from `pascal_parser.c` to the new files.
- Updated `pascal_parser.h` to include the new header files.
- Updated the `CMakeLists.txt` file to include the new source files in the build.
- Verified that all tests still pass after the refactoring.
