+++ examples/pascal_parser/pascal_statement.h
@@ -1,0 +1,9 @@
+#ifndef PASCAL_STATEMENT_H
+#define PASCAL_STATEMENT_H
+
+#include "parser.h"
+#include "combinators.h"
+
+void init_pascal_statement_parser(combinator_t** p);
+
+#endif // PASCAL_STATEMENT_H
